Sotre items in one swift file
