+++ FitPetner/code/GameViewController.swift
@@ -63,8 +63,8 @@ class GameViewController: UIViewController, ARSessionDelegate, ARSCNViewDelegate
         let tap = UITapGestureRecognizer(target: self, action: #selector(handleTap(_:)))
         tap.numberOfTouchesRequired = 1
         self.scnView.addGestureRecognizer(tap)
-
-        itemAppear()
+        
+        foodAppear()
         coinAppear()
         
         showUI()
@@ -247,109 +247,22 @@ class GameViewController: UIViewController, ARSessionDelegate, ARSCNViewDelegate
         // Dispose of any resources that can be recreated.
     }
     
-    // Let food appear, depending on time
-    func itemAppear(){
+    func foodAppear(){
         var itemToAppear: SCNNode!
-        itemToAppear = loadItem()
-        scnView.pointOfView?.addChildNode(itemToAppear)
+        itemToAppear = itemGenerator.loadFood()
+        //scnView.pointOfView?.addChildNode(itemToAppear)
+        scnView.scene.rootNode.addChildNode(itemToAppear)
         
         // Particle effect
-        let bokehEmitter = createBokeh()
+        let bokehEmitter = itemGenerator.createBokeh()
         itemToAppear.addParticleSystem(bokehEmitter)
-        
-        /*
-         // 用 physics 做動畫，有重力問題
-        itemToAppear.physicsBody = SCNPhysicsBody(type: .dynamic, shape: nil)
-        
-        let force = SCNVector3(x: 0, y: 5 , z: 0)
-        let position = SCNVector3(x: 0.05, y: 0.05, z: 0.05)
-        itemToAppear.physicsBody?.applyForce(force, at: position, asImpulse: true)
- */
-    }
-    
-    // Load 3D models that will appear in game
-    func loadItem() -> SCNNode{
-        // var itemArray : [SCNNode] = [] // Choose not to store items in an array
-        
-        // Randomly choose item to appear
-        let numberOfItem = Int(3) // 新增物品時記得改這邊的數字
-        let randomNumber = Int(arc4random_uniform(UInt32(numberOfItem))+1)
-        var itemLoaded: SCNNode!
-        
-        if (randomNumber == 1) {
-            // Load model
-            let scene = SCNScene( named: "art.scnassets/apple.dae")!
-            itemLoaded = scene.rootNode.childNode(withName: "apple", recursively: true)
-            itemLoaded.name = "FOOD_apple"
-            itemLoaded.scale = SCNVector3(x:0.02, y:0.02, z:0.02)
-        } else if (randomNumber == 2) {
-            // Load model
-            let scene = SCNScene( named: "art.scnassets/boletus.dae")!
-            itemLoaded = scene.rootNode.childNode(withName: "boletus", recursively: true)
-            itemLoaded.name = "FOOD_boletus"
-            itemLoaded.scale = SCNVector3(x:0.02, y:0.02, z:0.02)
-        } else if (randomNumber == 3) {
-            // Load model
-            let scene = SCNScene( named: "art.scnassets/raw_meat.dae")!
-            itemLoaded = scene.rootNode.childNode(withName: "Raw_meat", recursively: true)
-            itemLoaded.name = "FOOD_rawmeat"
-            itemLoaded.scale = SCNVector3(x: 0.25, y: 0.25, z: 0.25)
-        }
-        
-        itemLoaded.position = SCNVector3(x:0, y:-3, z:-10)
-        
-        let physicsBody = SCNPhysicsBody(
-            type: .kinematic,
-            shape: SCNPhysicsShape(geometry: SCNSphere(radius: 0.1))
-        )
-        itemLoaded.physicsBody = physicsBody
-        
-        return itemLoaded
     }
     
     func coinAppear(){
         var coinToAppear: SCNNode!
-        coinToAppear = loadCoin()
-        scnView.pointOfView?.addChildNode(coinToAppear)
-    }
-    
-    // Load Coins that will appear in game
-    func loadCoin() -> SCNNode{
-        // Randomly choose a type of coin to appear
-        let numberOfCoin = Int(2) // 新增物品時記得改這邊的數字
-        let randomNumber = Int(arc4random_uniform(UInt32(numberOfCoin))+1)
-        var coinLoaded: SCNNode!
-        
-        if (randomNumber == 1) {
-            // Load model
-            let scene = SCNScene( named: "art.scnassets/coin.dae")!
-            coinLoaded = scene.rootNode.childNode(withName: "coin", recursively: true)
-            coinLoaded.scale = SCNVector3(x:4, y:4, z:4)
-        } else if (randomNumber == 2) {
-            // Load model
-            let scene = SCNScene( named: "art.scnassets/coin2.dae")!
-            coinLoaded = scene.rootNode.childNode(withName: "coin2", recursively: true)
-            coinLoaded.scale = SCNVector3(x:50, y:50, z:50)
-        }
-        
-        coinLoaded.position = SCNVector3(x:-1, y:0, z:-10)
-        coinLoaded.name = "COIN"
-        
-        let physicsBody = SCNPhysicsBody(
-            type: .kinematic,
-            shape: SCNPhysicsShape(geometry: SCNSphere(radius: 0.1))
-        )
-        coinLoaded.physicsBody = physicsBody
-        
-        return coinLoaded
-    }
-    
-    // Particle effect for items appearing
-    func createBokeh() -> SCNParticleSystem {
-        let bokeh = SCNParticleSystem(named: "bokeh.scnp", inDirectory: nil)!
-        bokeh.particleColor = UIColor.white
-        bokeh.emitterShape = SCNGeometry()
-        return bokeh
+        coinToAppear = itemGenerator.loadCoin()
+        //scnView.pointOfView?.addChildNode(coinToAppear)
+        scnView.scene.rootNode.addChildNode(coinToAppear)
     }
     
     // Handle Tap on coin and food
@@ -377,7 +290,7 @@ class GameViewController: UIViewController, ARSessionDelegate, ARSCNViewDelegate
                 self.updateProgress()
                 
                 DispatchQueue.main.asyncAfter(deadline: .now() + 3.0, execute: {
-                    self.itemAppear()
+                    self.foodAppear()
                 })
             } else if resultNode.name == "COIN" {
                 resultNode.removeFromParentNode()
